requirements.txt
llama-index-core
llama-index-readers-file
gradio
python-dotenv
httpx